% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apparams_Bfabric}
\alias{apparams_Bfabric}
\title{get apparams from bfabric executable}
\usage{
apparams_Bfabric(yml)
}
\arguments{
\item{yml}{parsed YAML configuration list from B-Fabric}
}
\description{
get apparams from bfabric executable
}
