% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_helpers.R
\name{copy_SAINT_express}
\alias{copy_SAINT_express}
\title{copy Markdown and runscript for DIANN diann-output.tsv}
\usage{
copy_SAINT_express(workdir = getwd(), run_script = FALSE)
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}

\item{run_script}{if TRUE, also copy the R run scripts}
}
\value{
return value from `prolfqua::script_copy_helper_vec()`.
}
\description{
copy Markdown and runscript for DIANN diann-output.tsv
}
\examples{
copy_SAINT_express(workdir = tempdir(), run_script = FALSE)
}
