% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_DIANN_output.R
\name{dataset_set_factors_deprecated}
\alias{dataset_set_factors_deprecated}
\title{set factors and sample names columns}
\usage{
dataset_set_factors_deprecated(atable, msdata, repeated = TRUE, SAINT = FALSE)
}
\arguments{
\item{atable}{an AnalysisConfiguration object}

\item{msdata}{data.frame with annotation columns}

\item{repeated}{logical, look for subject/BioReplicate column}

\item{SAINT}{logical, use Bait_ instead of Group_ as factor name}
}
\value{
list with updated `atable` and `msdata`.
}
\description{
set factors and sample names columns
}
\examples{

dataset_csv_examples <- system.file("application/dataset_csv", package = "prolfquapp")
files <- dir(dataset_csv_examples, pattern = "*.csv")

res <- list()
for (i in seq_along(files)) {
  res[[i]] <- readr::read_csv(file.path(dataset_csv_examples, files[i]),
    show_col_types = FALSE)
}

for (i in seq_along(res)) {
  atable <- prolfqua::AnalysisConfiguration$new()
  atable$hierarchy[["protein_Id"]] <- c("Protein")
  atable$hierarchy[["peptide_Id"]] <- c("Peptide")
  tmp <- prolfquasaint::dataset_set_factors_deprecated(atable, res[[i]])
  cat(i, " : ", length(tmp$atable$factors), "factors : ",
    paste(tmp$atable$factors, collapse = "; "), "\n")
}

}
