% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_files_DIANN}
\alias{get_files_DIANN}
\title{find DIANN files}
\usage{
get_files_DIANN(path)
}
\arguments{
\item{path}{directory path to search for DIANN output files}
}
\value{
list with report TSV and FASTA file paths.
}
\description{
find DIANN files
}
\examples{
td <- tempdir()
writeLines("protein", file.path(td, "database.fasta"))
writeLines("report", file.path(td, "report.tsv"))
get_files_DIANN(td)
}
