% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_exp}
\alias{normalize_exp}
\title{normalize and then exponentiate data.}
\usage{
normalize_exp(lfqdataProt, normalization = c("vsn", "robscale", "none"))
}
\description{
normalize and then exponentiate data.
}
