% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_DIANN_output.R
\name{read_DIANN_output}
\alias{read_DIANN_output}
\title{DIANN helper functions}
\usage{
read_DIANN_output(
  diann.path,
  fasta.file,
  nrPeptides = 2,
  q_value = 0.01,
  isUniprot = TRUE,
  rev = "REV_"
)
}
\arguments{
\item{diann.path}{path to diann-output.tsv}

\item{fasta.file}{path to fasta file}

\item{nrPeptides}{peptide threshold}

\item{q_value}{q value threshold}

\item{isUniprot}{logical, is the fasta file from UniProt}

\item{rev}{prefix for reversed/decoy entries}
}
\value{
This function now errors and points users to prolfquapp.
}
\description{
DIANN helper functions
}
\examples{
try(read_DIANN_output("report.tsv", "database.fasta"), silent = TRUE)
}
