% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_force}
\alias{transform_force}
\title{force data transformation}
\usage{
transform_force(lfqdataProt, transformation = c("sqrt", "log2", "none"))
}
\arguments{
\item{lfqdataProt}{an LFQData object}

\item{transformation}{transformation method: "sqrt", "log2", or "none"}
}
\value{
transformed LFQData object.
}
\description{
force data transformation
}
\examples{
obj <- list()
transform_force(obj, transformation = "none")
}
